demo-supplychain-ai/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.0.0
django>=5.0.0
